Support labels in job dependencies
It is a common pattern to depend on the same set of multiple jobs from
multiple places.  Additionally with orchestration tools this set of job
may be dynamic based on configuration.  In kubernetes we can encapsulate
this set of jobs via shared labels.  This patchset then adds support to
depend on such a set of jobs via labels.

This moves the DEPENDENCY_JOBS env var to use JSON, similar to
DEPENDENCY_POD, in order to support labels, which is a breaking change.